+++ packages/desktop_autopaste_platform_interface/lib/desktop_autopaste_platform_interface.dart
@@ -1,0 +1,6 @@
+library;
+
+export 'src/desktop_autopaste_platform.dart';
+export 'src/focused_text_edit_operation.dart';
+export 'src/focused_text_field_context.dart';
+export 'src/paste_shortcut.dart';